+++ packages/components/package.json
@@ -12,7 +12,7 @@
   "license": "GPL-3.0",
   "scripts": {
     "test": "jest",
-    "lint": "eslint . --max-warnings=94"
+    "lint": "eslint . --max-warnings=971"
   },
   "jest": {
     "testRegex": ".*Test.js$",
