+++ packages/yoast-components/composites/Plugin/Shared/components/ButtonSection.js
@@ -6,7 +6,7 @@ import {
 	StyledIconsButton,
 	StyledContainer,
 	StyledContainerTopLevel,
-} from "./Collapsible";
+} from "@yoast/components";
 import { SectionTitle } from "@yoast/components";
 
 const StyledHeading = wrapInHeading( StyledIconsButton, { level: 2, fontSize: "1rem", fontWeight: "normal" } );
